Move fs.sshfs to it's own namespace
