+++ setup.cfg
@@ -29,7 +29,7 @@ classifiers =
 zip_safe = true
 include_package_data = true
 python_requires = >= 2.7, != 3.0.*, != 3.1.*, != 3.2.*, != 3.3.*
-packages = fs.sshfs, fs.opener
+packages = sshfs
 test_suite = tests
 setup_requires = setuptools
 install_requires =
@@ -55,8 +55,8 @@ test =
 
 [options.entry_points]
 fs.opener =
-	ssh  = fs.opener.sshfs:SSHOpener
-	sftp = fs.opener.sshfs:SSHOpener
+	ssh  = sshfs.opener:SSHOpener
+	sftp = sshfs.opener:SSHOpener
 
 [coverage:report]
 show_missing = true
@@ -73,7 +73,7 @@ exclude_lines =
 verbosity = 2
 with-coverage = true
 cover-xml = true
-cover-package = fs.sshfs, fs.opener.sshfs
+cover-package = sshfs._sshfs, sshfs.opener
 with-doctest = true
 doctest-extension = .rst
 rednose = true
