+++ sshfs/_sshfs.py
@@ -13,17 +13,17 @@ import sys
 import six
 import paramiko
 
-from .. import errors
-from ..base import FS
-from ..info import Info
-from ..enums import ResourceType
-from ..iotools import RawWrapper
-from ..path import basename
-from ..permissions import Permissions
-from ..osfs import OSFS
-from ..mode import Mode
+from fs import errors
+from fs.base import FS
+from fs.info import Info
+from fs.enums import ResourceType
+from fs.iotools import RawWrapper
+from fs.path import basename
+from fs.permissions import Permissions
+from fs.osfs import OSFS
+from fs.mode import Mode
 
-from .file import SSHFile
+from ._file import SSHFile
 from .error_tools import convert_sshfs_errors
