+++ sshfs/opener.py
@@ -6,25 +6,9 @@ import configparser
 
 import six
 
-from .base import Opener
-from ..subfs import ClosingSubFS
-from ..errors import FSError, CreateFailed
-
-__license__ = "LGPL-2.1+"
-__copyright__ = "Copyright (c) 2017 Martin Larralde"
-__author__ = "Martin Larralde <martin.larralde@ens-cachan.fr>"
-__version__ = 'dev'
-
-
-# Dynamically get the version of the main module
-try:
-    _name = __name__.replace('.opener', '')
-    import pkg_resources
-    __version__ = pkg_resources.get_distribution(_name).version
-except Exception: # pragma: no cover
-    pkg_resources = None
-finally:
-    del pkg_resources
+from fs.opener.base import Opener
+from fs.subfs import ClosingSubFS
+from fs.errors import FSError, CreateFailed
 
 
 class SSHOpener(Opener):
@@ -32,7 +16,7 @@ class SSHOpener(Opener):
 
     @staticmethod
     def open_fs(fs_url, parse_result, writeable, create, cwd):
-        from ..sshfs import SSHFS
+        from ._sshfs import SSHFS
         ssh_host, _, dir_path = parse_result.resource.partition('/')
         ssh_host, _, ssh_port = ssh_host.partition(':')
         ssh_port = int(ssh_port) if ssh_port.isdigit() else 22
